backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
ffmpeg-python
numpy
scipy
python-dotenv
aiofiles
google-auth
requests
